Destroyed Utils by splitting it out into FloatBytePacker and NetworkManager.
