+++ Assets/Mirror/Runtime/FloatBytePacker.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: afd3cca6a786d4208b1d0f7f2b168901
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
